ui: Move Poster components to `poster` folder
gotta organize - gonna be more in a second
